requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]>=3.1
python-decouple==3.8
Pillow==10.4.0

# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Redis & Caching
redis==5.0.1
django-redis==5.4.0

# Celery for Background Tasks
celery==5.3.4

# Async Support
asgiref==3.8.1
uvicorn==0.27.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality & Pre-commit
pre-commit==3.6.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
bandit==1.7.6
